x/rate/client/cli/query_era_exchange_rate.go
package cli

import (
    "context"
	"strconv"

	"github.com/spf13/cobra"

	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
    "github.com/stafiprotocol/stafihub/x/rate/types"
)

func CmdShowEraExchangeRate() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-era-exchange-rate [denom] [era]",
		Short: "shows a EraExchangeRate",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) (err error) {
			argDenom := args[0]

			argEra, err := strconv.ParseUint(args[1], 10, 64)
			if err != nil {
				return err
			}

            clientCtx := client.GetClientContextFromCmd(cmd)

            queryClient := types.NewQueryClient(clientCtx)

            params := &types.QueryGetEraExchangeRateRequest{
				argDenom,
				uint32(argEra),
            }

            res, err := queryClient.EraExchangeRate(context.Background(), params)
            if err != nil {
                return err
            }

            return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

    return cmd
}
